requirements.txt
PyPDF2>=3.0.0
openai>=1.0.0
python-dotenv>=1.0.0
tenacity>=8.0.0
typing-extensions>=4.0.0
pathlib>=1.0.0
logging>=0.5.0 